+++ libiclight/args.py
@@ -1,0 +1,224 @@
+from typing import Optional
+import numpy as np
+import torch
+from enum import Enum
+from pydantic import BaseModel, validator
+
+from modules import scripts
+from modules.api import api
+from modules.processing import (
+    StableDiffusionProcessing,
+    StableDiffusionProcessingTxt2Img,
+)
+
+from libiclight.utils import (
+    BriarmbgService,
+    align_dim_latent,
+    make_masked_area_grey,
+    resize_and_center_crop,
+    ldm_numpy2pytorch,
+)
+
+
+class BGSourceFC(Enum):
+    """BGSource for FC model."""
+
+    NONE = "None"
+    LEFT = "Left Light"
+    RIGHT = "Right Light"
+    TOP = "Top Light"
+    BOTTOM = "Bottom Light"
+    CUSTOM = "Custom LightMap"
+
+    def get_bg(
+        self,
+        image_width: int,
+        image_height: int,
+        **kwargs,
+    ) -> np.ndarray:
+        bg_source = self
+        if bg_source == BGSourceFC.NONE:
+            pass
+        elif bg_source == BGSourceFC.LEFT:
+            gradient = np.linspace(255, 0, image_width)
+            image = np.tile(gradient, (image_height, 1))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFC.RIGHT:
+            gradient = np.linspace(0, 255, image_width)
+            image = np.tile(gradient, (image_height, 1))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFC.TOP:
+            gradient = np.linspace(255, 0, image_height)[:, None]
+            image = np.tile(gradient, (1, image_width))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFC.BOTTOM:
+            gradient = np.linspace(0, 255, image_height)[:, None]
+            image = np.tile(gradient, (1, image_width))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        else:
+            raise "Wrong initial latent!"
+
+        return input_bg
+
+
+class BGSourceFBC(Enum):
+    """BGSource for FBC model."""
+
+    UPLOAD = "Use Background Image"
+    UPLOAD_FLIP = "Use Flipped Background Image"
+    LEFT = "Left Light"
+    RIGHT = "Right Light"
+    TOP = "Top Light"
+    BOTTOM = "Bottom Light"
+    GREY = "Ambient"
+
+    def get_bg(
+        self,
+        image_width: int,
+        image_height: int,
+        uploaded_bg: Optional[np.ndarray] = None,
+    ) -> np.ndarray:
+        bg_source = self
+        if bg_source == BGSourceFBC.UPLOAD:
+            assert uploaded_bg is not None
+            input_bg = uploaded_bg
+        elif bg_source == BGSourceFBC.UPLOAD_FLIP:
+            input_bg = np.fliplr(uploaded_bg)
+        elif bg_source == BGSourceFBC.GREY:
+            input_bg = (
+                np.zeros(shape=(image_height, image_width, 3), dtype=np.uint8) + 64
+            )
+        elif bg_source == BGSourceFBC.LEFT:
+            gradient = np.linspace(224, 32, image_width)
+            image = np.tile(gradient, (image_height, 1))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFBC.RIGHT:
+            gradient = np.linspace(32, 224, image_width)
+            image = np.tile(gradient, (image_height, 1))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFBC.TOP:
+            gradient = np.linspace(224, 32, image_height)[:, None]
+            image = np.tile(gradient, (1, image_width))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        elif bg_source == BGSourceFBC.BOTTOM:
+            gradient = np.linspace(32, 224, image_height)[:, None]
+            image = np.tile(gradient, (1, image_width))
+            input_bg = np.stack((image,) * 3, axis=-1).astype(np.uint8)
+        else:
+            raise "Wrong background source!"
+        return input_bg
+
+
+class ModelType(Enum):
+    FC = "FC"
+    FBC = "FBC"
+
+    @property
+    def model_name(self) -> str:
+        if self == ModelType.FC:
+            return "iclight_sd15_fc_unet_ldm.safetensors"
+        else:
+            assert self == ModelType.FBC
+            return "iclight_sd15_fbc_unet_ldm.safetensors"
+
+
+class ICLightArgs(BaseModel):
+    enabled: bool = False
+    model_type: ModelType = ModelType.FC
+    input_fg: Optional[np.ndarray] = None
+    uploaded_bg: Optional[np.ndarray] = None
+    bg_source_fc: BGSourceFC = BGSourceFC.NONE
+    bg_source_fbc: BGSourceFBC = BGSourceFBC.UPLOAD
+    remove_bg: bool = True
+
+    @classmethod
+    def cls_decode_base64(cls, base64string: str) -> np.ndarray:
+        return np.array(api.decode_base64_to_image(base64string)).astype("uint8")
+
+    @validator("input_fg", pre=True)
+    def parse_input_fg(cls, value) -> np.ndarray:
+        if isinstance(value, str):
+            return cls.cls_decode_base64(value)
+        assert isinstance(value, np.ndarray) or value is None
+        return value
+
+    @validator("uploaded_bg", pre=True)
+    def parse_input_bg(cls, value) -> np.ndarray:
+        if isinstance(value, str):
+            return cls.cls_decode_base64(value)
+        assert isinstance(value, np.ndarray) or value is None
+        return value
+
+    @classmethod
+    def fetch_from(cls, p: StableDiffusionProcessing):
+        script_runner: scripts.ScriptRunner = p.scripts
+        ic_light_script: scripts.Script = [
+            script
+            for script in script_runner.alwayson_scripts
+            if script.title() == "IC Light"
+        ][0]
+        args = p.script_args[ic_light_script.args_from : ic_light_script.args_to]
+        assert len(args) == 1
+        return ICLightArgs(**args[0])
+
+    class Config:
+        arbitrary_types_allowed = True
+
+    def get_c_concat(
+        self,
+        processed_fg: np.ndarray,  # fg with bg removed.
+        vae,
+        p: StableDiffusionProcessing,
+        device: torch.device,
+    ) -> dict:
+        is_hr_pass = getattr(p, "is_hr_pass", False)
+        if is_hr_pass:
+            assert isinstance(p, StableDiffusionProcessingTxt2Img)
+            # TODO: Move the calculation to Forge main repo.
+            if p.hr_resize_x == 0 and p.hr_resize_y == 0:
+                hr_x = int(p.width * p.hr_scale)
+                hr_y = int(p.height * p.hr_scale)
+            else:
+                hr_y, hr_x = p.hr_resize_y, p.hr_resize_x
+            image_width = align_dim_latent(hr_x)
+            image_height = align_dim_latent(hr_y)
+        else:
+            image_width = p.width
+            image_height = p.height
+
+        fg = resize_and_center_crop(processed_fg, image_width, image_height)
+        if self.model_type == ModelType.FC:
+            np_concat = [fg]
+        else:
+            assert self.model_type == ModelType.FBC
+            bg = resize_and_center_crop(
+                self.bg_source_fbc.get_bg(image_width, image_height, self.uploaded_bg),
+                image_width,
+                image_height,
+            )
+            np_concat = [fg, bg]
+
+        concat_conds = ldm_numpy2pytorch(np.stack(np_concat, axis=0)).to(
+            device=device, dtype=torch.float16
+        )
+
+        # Optional: Use mode instead of sample from VAE output.
+        # vae.first_stage_model.regularization.sample = False
+        latent_concat_conds = vae.encode(concat_conds)
+        return {"samples": latent_concat_conds}
+
+    def get_input_rgb(self, device: torch.device) -> np.ndarray:
+        if self.remove_bg:
+            input_fg = self.input_fg[..., :3]
+            alpha = BriarmbgService().run_rmbg(img=input_fg, device=device)
+            input_rgb: np.ndarray = make_masked_area_grey(input_fg, alpha)
+        else:
+            if self.input_fg.shape[-1] == 4:
+                input_rgb = make_masked_area_grey(
+                    self.input_fg[..., :3],
+                    self.input_fg[..., 3:].astype(np.float32) / 255.0,
+                )
+            else:
+                input_rgb = self.input_fg
+        assert input_rgb.shape[-1] == 3, "Input RGB should have 3 channels."
+        return input_rgb
